chore: populate empty test class and rename classes
