+++ e2eTests/src/test/java/no/elixir/e2eTests/features/DownloadTest.java
@@ -1,105 +1,106 @@
 package no.elixir.e2eTests.features;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
 
-import java.util.Collection;
-import java.util.Optional;
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.StringWriter;
+import java.security.KeyPair;
 import kong.unirest.HttpResponse;
-import kong.unirest.JsonNode;
 import kong.unirest.Unirest;
+import no.elixir.crypt4gh.stream.Crypt4GHInputStream;
 import no.elixir.e2eTests.constants.Strings;
 import no.elixir.e2eTests.core.E2EState;
+import no.elixir.e2eTests.utils.CommonUtils;
 import no.elixir.e2eTests.utils.TokenUtils;
+import org.apache.commons.codec.binary.Hex;
+import org.apache.commons.codec.digest.DigestUtils;
+import org.apache.commons.io.IOUtils;
+import org.skyscreamer.jsonassert.JSONAssert;
 
 public class DownloadTest {
 
-  public static void testDownloadDatasetUsingExportRequestAndVerifyResults() throws Exception {
-    String accessToken = TokenUtils.generateVisaToken(E2EState.datasetId);
-    E2EState.log.info(E2EState.encFile.getName());
-
-    HttpResponse<JsonNode> exportRequestRes = makeExportRequest(accessToken);
-    assertEquals(200, exportRequestRes.getStatus());
-
-    E2EState.log.info("Export request response: {}", exportRequestRes.getBody());
-    HttpResponse<FileListingResponse> listFilesRes = checkFilesWithRetry(accessToken);
-    assertNotNull(listFilesRes);
-    assertEquals(200, listFilesRes.getStatus());
-
-    E2EState.log.info("List user outbox request response: {}", listFilesRes.getBody());
-    assertFalse(listFilesRes.getBody().files.isEmpty());
-
-    Optional<TsdFile> first = listFilesRes.getBody().files.stream().findFirst();
-    assertTrue(first.isPresent());
-    assertEquals(E2EState.encFile.getName(), first.get().fileName);
-  }
-
-  private static HttpResponse<JsonNode> makeExportRequest(String accessToken) throws Exception {
-    E2EState.log.info("Preparing to make export request");
-    String exportReqUrl = buildExportUrl();
-    String payload = buildExportPayload(accessToken);
-    E2EState.log.info("Export request payload: {}", payload);
-    return Unirest.post(exportReqUrl)
-        .body(payload)
-        .contentType("application/json")
-        .basicAuth(E2EState.env.getProxyAdminUsername(), E2EState.env.getProxyAdminPassword())
-        .asJson();
-  }
-
-  private static HttpResponse<FileListingResponse> checkFilesWithRetry(String accessToken)
-      throws Exception {
-    String listFilesEndpoint = buildListFilesUrl();
-    final int MAX_RETRIES = 5;
-    for (int i = 1; i <= MAX_RETRIES; i++) {
-      HttpResponse<FileListingResponse> res =
-          Unirest.get(listFilesEndpoint)
-              .socketTimeout(300000)
-              .header("Proxy-Authorization", "Bearer " + accessToken)
-              .asObject(FileListingResponse.class);
-      E2EState.log.info("List files request response: {}", res);
-      // if something went wrong with this request
-      if (200 != res.getStatus()) {
-        return res;
-      }
-      FileListingResponse body = res.getBody();
-      // Check if files array exists and is not empty
-      if (!body.files.isEmpty()) {
-        E2EState.log.info("Files found on attempt {}", i);
-        return res;
-      }
-      if (i < MAX_RETRIES) {
-        E2EState.log.info("Files not found, waiting 1 second before attempt {}", i + 1);
-        Thread.sleep(1000);
-      }
+  /** Test and check that what we get out match the original inserted data at the top. */
+  public static void downloadDatasetAndVerifyResults() throws Exception {
+    String token = TokenUtils.generateVisaToken(E2EState.datasetId);
+    E2EState.log.info("Visa JWT token when downloading: {}", token);
+    String datasets =
+        Unirest.get(
+                String.format(
+                    "https://%s:%s/metadata/datasets",
+                    E2EState.env.getSdaDoaHost(), E2EState.env.getSdaDoaPort()))
+            .header("Authorization", "Bearer " + token)
+            .asString()
+            .getBody();
+    assertEquals(String.format("[\"%s\"]", E2EState.datasetId).strip(), datasets.strip());
+    // Meta data check
+    String expected =
+        CommonUtils.toCompactJson(
+            String.format(
+                    Strings.EXPECTED_DOWNLOAD_METADATA,
+                    E2EState.stableId,
+                    E2EState.datasetId,
+                    E2EState.encFile.getName(),
+                    E2EState.archivePath,
+                    E2EState.rawSHA256Checksum)
+                .strip());
+    String actual =
+        CommonUtils.toCompactJson(
+            Unirest.get(
+                    String.format(
+                        "https://%s:%s/metadata/datasets/%s/files",
+                        E2EState.env.getSdaDoaHost(),
+                        E2EState.env.getSdaDoaPort(),
+                        E2EState.datasetId))
+                .header("Authorization", "Bearer " + token)
+                .asString()
+                .getBody()
+                .strip());
+    E2EState.log.info("Expected: {}", expected);
+    E2EState.log.info("Actual: {}", actual);
+    JSONAssert.assertEquals(expected, actual, false);
+    // Fetch the non-encrypted file
+    HttpResponse<byte[]> response =
+        Unirest.get(
+                String.format(
+                    "https://%s:%s/files/%s",
+                    E2EState.env.getSdaDoaHost(), E2EState.env.getSdaDoaPort(), E2EState.stableId))
+            .header("Authorization", "Bearer " + token)
+            .asBytes();
+    if (response.getStatus() == 200) { // Check if the response is OK
+      byte[] file = response.getBody();
+      String obtainedChecksum = Hex.encodeHexString(DigestUtils.sha256(file));
+      assertEquals(E2EState.rawSHA256Checksum, obtainedChecksum);
+    } else {
+      fail("Failed to fetch the file. Status: " + response.getStatus());
     }
-    E2EState.log.warn("No files found after 5 attempts");
-    return null;
+    // Fetch the encrypted file
+    KeyPair recipientKeyPair = E2EState.keyUtils.generateKeyPair();
+    StringWriter stringWriter = new StringWriter();
+    E2EState.keyUtils.writeCrypt4GHKey(stringWriter, recipientKeyPair.getPublic(), null);
+    String key = stringWriter.toString();
+    HttpResponse<byte[]> encFileRes =
+        Unirest.get(
+                String.format(
+                    "https://%s:%s/files/%s?destinationFormat=CRYPT4GH",
+                    E2EState.env.getSdaDoaHost(), E2EState.env.getSdaDoaPort(), E2EState.stableId))
+            .header("Authorization", "Bearer " + token)
+            .header("Public-Key", key)
+            .asBytes();
+    if (encFileRes.getStatus() == 200) { // Check if the response is OK
+      byte[] file = encFileRes.getBody();
+      ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+      try (ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(file);
+          Crypt4GHInputStream crypt4GHInputStream =
+              new Crypt4GHInputStream(byteArrayInputStream, recipientKeyPair.getPrivate())) {
+        IOUtils.copyLarge(crypt4GHInputStream, byteArrayOutputStream);
+      }
+      String obtainedChecksum =
+          Hex.encodeHexString(DigestUtils.sha256(byteArrayOutputStream.toByteArray()));
+      assertEquals(E2EState.rawSHA256Checksum, obtainedChecksum);
+    } else {
+      fail("Failed to fetch the encrypted file. Status: " + response.getStatus());
+    }
   }
-
-  private static String buildExportUrl() {
-    return String.format(
-        "https://%s:%s/export", E2EState.env.getProxyHost(), E2EState.env.getProxyPort());
-  }
-
-  private static String buildListFilesUrl() {
-    return String.format(
-        "https://%s:%s/files?inbox=false",
-        E2EState.env.getProxyHost(), E2EState.env.getProxyPort());
-  }
-
-  private static String buildExportPayload(String accessToken) throws Exception {
-    return Strings.EXPORT_REQ_BODY.formatted(
-        E2EState.datasetId, accessToken, TokenUtils.encodedPublicKey(), "DATASET_ID");
-  }
-
-  record FileListingResponse(Collection<TsdFile> files, String page) {}
-
-  record TsdFile(
-      String fileName,
-      Long size,
-      String modifiedDate,
-      String href,
-      Boolean exportable,
-      String reason,
-      String mimeType,
-      String owner) {}
 }
